+++ src/config/configuration.ts
@@ -1,9 +1,4 @@
 export default () => ({
   port: process.env.PORT,
-
-  // DB connection comming soon
-  // database: {
-  //   host: process.env.DATABASE_HOST,
-  //   port: parseInt(process.env.DATABASE_PORT, 10) || 5432,
-  // },
+  secret: process.env.JWT_SECRET,
 });
