--- src/constatns/jwt.ts
@@ -1,3 +1,0 @@
-export const jwtConstants = {
-  secret: `${process.env.JWT_SECRET}`,
-};
