+++ src/domains/database/database.module.ts
@@ -1,0 +1,8 @@
+import { Module } from '@nestjs/common';
+import { databaseProviders } from './database.repository';
+
+@Module({
+  providers: [...databaseProviders],
+  exports: [...databaseProviders],
+})
+export class DatabaseModule {}
