+++ src/domains/seats/seats.repository.ts
@@ -1,0 +1,10 @@
+import { DataSource } from 'typeorm';
+import { Seats } from '@entities/seats.entity';
+
+export const seatsProviders = [
+  {
+    provide: 'SEATS_REPOSITORY',
+    useFactory: (dataSource: DataSource) => dataSource.getRepository(Seats),
+    inject: ['DATA_SOURCE'],
+  },
+];
