+++ src/domains/users/users.module.ts
@@ -1,0 +1,12 @@
+import { Module } from '@nestjs/common';
+
+import { UsersService } from './users.service';
+import { UsersController } from './users.controller';
+import { usersProviders } from './users.repository';
+
+@Module({
+  controllers: [UsersController],
+  providers: [UsersService, ...usersProviders],
+  exports: [UsersService],
+})
+export class UsersModule {}
