+++ src/domains/users/users.repository.ts
@@ -1,0 +1,10 @@
+import { DataSource } from 'typeorm';
+import { User } from '@entities/user.entity';
+
+export const usersProviders = [
+  {
+    provide: 'USER_REPOSITORY',
+    useFactory: (dataSource: DataSource) => dataSource.getRepository(User),
+    inject: ['DATA_SOURCE'],
+  },
+];
