+++ src/entities/index.ts
@@ -1,0 +1,3 @@
+export * from './auth.entity';
+export * from './seats.entity';
+export * from './user.entity';
