--- src/health/health.controller.spec.ts
@@ -1,44 +1,0 @@
-import { Test, TestingModule } from '@nestjs/testing';
-import { HealthController } from './health.controller';
-import { HealthCheckService, HttpHealthIndicator } from '@nestjs/terminus';
-import { ConfigService } from '@nestjs/config';
-
-describe('HealthController', () => {
-  let controller: HealthController;
-
-  const mockHealthCheckService = {};
-  const mockHttpHealthIndicator = {};
-  const mockConfigService = {};
-
-  beforeEach(async () => {
-    const module: TestingModule = await Test.createTestingModule({
-      controllers: [HealthController],
-      providers: [
-        {
-          provide: HealthCheckService,
-          useValue: mockHealthCheckService,
-        },
-        {
-          provide: HttpHealthIndicator,
-          useValue: mockHttpHealthIndicator,
-        },
-        {
-          provide: ConfigService,
-          useValue: mockConfigService,
-        },
-      ],
-    }).compile();
-
-    controller = module.get<HealthController>(HealthController);
-  });
-
-  it('should be defined', () => {
-    expect(controller).toBeDefined();
-  });
-
-  describe('health', () => {
-    it('should be health!"', () => {
-      expect(controller.ping()).toBe('health');
-    });
-  });
-});
