--- src/seats/seats.providers.ts
@@ -1,10 +1,0 @@
-import { DataSource } from 'typeorm';
-import { Seats } from '@/entities/seats.entity';
-
-export const seatsProviders = [
-  {
-    provide: 'SEATS_REPOSITORY',
-    useFactory: (dataSource: DataSource) => dataSource.getRepository(Seats),
-    inject: ['DATA_SOURCE'],
-  },
-];
