--- src/users/users.module.ts
@@ -1,14 +1,0 @@
-import { Module } from '@nestjs/common';
-import { DatabaseModule } from '@/database/database.module';
-
-import { UsersService } from './users.service';
-import { UsersController } from './users.controller';
-import { usersProviders } from './users.providers';
-
-@Module({
-  // imports: [DatabaseModule],
-  controllers: [UsersController],
-  providers: [UsersService, ...usersProviders],
-  exports: [UsersService],
-})
-export class UsersModule {}
